* fixed moar *nix stuff
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@1695 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/GuildHandler.cpp
@@ -245,8 +245,8 @@ void WorldSession::HandleGuildInfo(WorldPacket & recv_data)
 	data << pGuild->GetCreatedYear();
 	data << pGuild->GetCreatedMonth();
 	data << pGuild->GetCreatedDay();
-	data << pGuild->GetGuildMembersCount();
-	data << pGuild->GetGuildMembersCount();//accountcount
+	data << uint32(pGuild->GetGuildMembersCount());
+	data << uint32(pGuild->GetGuildMembersCount());//accountcount
 
 	SendPacket(&data);
 }
